internals go to a separate header file
